refactor(web): eliminate Pages/ indirection, inline into route files
Move all page components directly into their TanStack route files so each
route is self-contained. Shared TemplatesPage (used by /templates and
/content-manager/templates) moves to src/components/templates-page.tsx.
Deletes the src/Pages/ directory entirely.
