+++ web/src/routes/templates.tsx
@@ -1,5 +1,5 @@
-import TemplatesPage from "@/Pages/templates";
-import { createFileRoute } from '@tanstack/react-router'
+import TemplatesPage from "@/components/templates-page";
+import { createFileRoute } from '@tanstack/react-router';
 
 
 export const Route = createFileRoute("/templates")({
